# internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for Tinylink",
        "title": "TinyLink API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
        "/tinylink/list": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves a list of Tinylinks for the authenticated user. This is protected route. Access tokey is needed",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tinylink"
                ],
                "summary": "List Tinylinks",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/tinylink.Tinylink"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/jsonutil.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/jsonutil.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "jsonutil.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "omitempty": {}
            }
        },
        "tinylink.Tinylink": {
            "type": "object",
            "properties": {
                "alias": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "domain": {
                    "type": "string"
                },
                "expiration": {
                    "type": "string"
                },
                "guest_id": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "private": {
                    "type": "boolean"
                },
                "updated_at": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "version": {
                    "type": "integer"
                }
            }
        }
    }
}